+++ flutter/lib/src/flutter_enricher_event_processor.dart
