+++ flutter/lib/src/sentry_flutter.dart
@@ -4,7 +4,7 @@ import 'package:flutter/services.dart';
 import 'package:package_info_plus/package_info_plus.dart';
 import 'package:sentry/sentry.dart';
 
-import 'flutter_enricher.dart';
+import 'flutter_enricher_event_processor.dart';
 import 'sentry_flutter_options.dart';
 
 import 'default_integrations.dart';
